move non-route functions to separate module
